src/requirements.txt
# requirements.txt
django>=4.0
djangorestframework

